students/k3341/Dushchenko_Daniil/Lr2/lab_2/requirements.txt
sqlmodel
psycopg2-binary
python-dotenv
requests
aiohttp
